+++ src/math_extra/angles.c
@@ -1,16 +1,16 @@
 #include "olafur.h"
 
-double	deg2rad(double deg)
+float	deg2rad(float deg)
 {
 	return (deg * PI_OVER_180);
 }
 
-double	get_angle_from_dir(t_vec2 dir)
+float	get_angle_from_dir(t_vec2 dir)
 {
-	return (atan2(dir.y, dir.x));
+	return (atan2f(dir.y, dir.x));
 }
 
-double	norm_angle(double angle)
+float	norm_angle(float angle)
 {
 	while (angle < 0)
 		angle += RAD_360;
@@ -19,16 +19,16 @@ double	norm_angle(double angle)
 	return (angle);
 }
 
-double	dot(t_vec2 a, t_vec2 b)
+float	dot(t_vec2 a, t_vec2 b)
 {
 	return (a.x * b.x + a.y * b.y);
 }
 
-t_vec2	rotate_vec2(t_vec2 v, double angle)
+t_vec2	rotate_vec2(t_vec2 v, float angle)
 {
 	t_vec2	rotated;
 
-	rotated.x = v.x * cos(angle) - v.y * sin(angle);
-	rotated.y = v.x * sin(angle) + v.y * cos(angle);
+	rotated.x = v.x * cosf(angle) - v.y * sinf(angle);
+	rotated.y = v.x * sinf(angle) + v.y * cosf(angle);
 	return (rotated);
 }
